Updated Takeoff class and GetCurrentWeather class
Changed the return type for the getWeather() method within the
GetCurrentWeather class. The weather variable within the Takeoff class
should be able to get the current weather as a String and  be used
within the takeoff() method. Errors still exist. I believe it's because
the Takeoff class isn't in the same package as the Flight, FlightList,
and GetCurrentWeather classes.